requirement.txt
streamlit
python-dotenv
google-generativeai
PyMuPDF